Prepare Assembly.cs, Module.cs and ParameterInfo.cs for move to shared partition.
